+++ packages/backend-defaults/src/CreateBackend.test.ts
@@ -16,16 +16,9 @@
 
 import {
   coreServices,
-  createBackendPlugin,
   createServiceFactory,
-  createServiceRef,
-  createSharedEnvironment,
 } from '@backstage/backend-plugin-api';
-import { mockServices } from '@backstage/backend-test-utils';
 import { createBackend } from './CreateBackend';
-
-const fooServiceRef = createServiceRef<string>({ id: 'foo', scope: 'root' });
-const barServiceRef = createServiceRef<string>({ id: 'bar', scope: 'root' });
 
 describe('createBackend', () => {
   it('should not throw when overriding a default service implementation', () => {
@@ -84,92 +77,5 @@ describe('createBackend', () => {
         ],
       }),
     ).toThrow('The core.pluginMetadata service cannot be overridden');
-  });
-
-  it('should throw if an unsupported InternalSharedEnvironment version is passed in', () => {
-    expect(() =>
-      createBackend({
-        env: {} as any,
-      }),
-    ).toThrow(
-      "Shared environment version 'undefined' is invalid or not supported",
-    );
-    expect(() =>
-      createBackend({
-        env: { version: {} } as any,
-      }),
-    ).toThrow(
-      "Shared environment version '[object Object]' is invalid or not supported",
-    );
-    expect(() =>
-      createBackend({
-        env: { version: 'v2' } as any,
-      }),
-    ).toThrow("Shared environment version 'v2' is invalid or not supported");
-  });
-
-  it('should prioritize services correctly', async () => {
-    const backend = createBackend({
-      env: createSharedEnvironment({
-        services: [
-          createServiceFactory({
-            service: coreServices.rootHttpRouter,
-            deps: {},
-            async factory() {
-              return {
-                use() {},
-              };
-            },
-          }),
-          mockServices.config.factory({
-            data: { root: 'root-env' },
-          }),
-          createServiceFactory({
-            service: fooServiceRef,
-            deps: {},
-            async factory() {
-              return 'foo-env';
-            },
-          }),
-          createServiceFactory({
-            service: barServiceRef,
-            deps: {},
-            async factory() {
-              return 'bar-env';
-            },
-          }),
-        ],
-      })(),
-      services: [
-        createServiceFactory({
-          service: fooServiceRef,
-          deps: {},
-          factory: async () => 'foo-backend',
-        }),
-      ],
-    });
-
-    expect.assertions(3);
-    backend.add(
-      createBackendPlugin({
-        pluginId: 'test',
-        register(reg) {
-          reg.registerInit({
-            deps: {
-              config: coreServices.config,
-              foo: fooServiceRef,
-              bar: barServiceRef,
-            },
-            async init({ config, foo, bar }) {
-              expect(config.get('root')).toBe('root-env');
-              expect(foo).toBe('foo-backend');
-              expect(bar).toBe('bar-env');
-            },
-          });
-        },
-      })(),
-    );
-
-    await backend.start();
   });
 });