+++ packages/cli/src/commands/index.ts
@@ -365,6 +365,8 @@ export function registerCommands(program: Command) {
       '--format <format>',
       'Format to print the schema in, either json or yaml [yaml]',
     )
+    .option('--merge', 'Print the config schemas merged', true)
+    .option('--no-merge', 'Print the config schemas not merged')
     .description('Print configuration schema')
     .action(lazy(() => import('./config/schema').then(m => m.default)));