+++ plugins/catalog-backend-module-github/src/module/catalogModuleGithubEntityProvider.test.ts
@@ -62,7 +62,7 @@ describe('catalogModuleGithubEntityProvider', () => {
     await startTestBackend({
       extensionPoints: [[catalogProcessingExtensionPoint, extensionPoint]],
       services: [
-        [coreServices.config, config],
+        [coreServices.rootConfig, config],
         [coreServices.logger, getVoidLogger()],
         [coreServices.scheduler, scheduler],
       ],