+++ plugins/catalog-backend-module-gitlab/src/lib/client.ts
@@ -30,6 +30,7 @@ export type CommonListOptions = {
 };
 
 interface ListProjectOptions extends CommonListOptions {
+  archived?: boolean;
   group?: string;
 }