+++ plugins/catalog-backend-module-gitlab/src/module/catalogModuleGitlabDiscoveryEntityProvider.test.ts
@@ -74,7 +74,7 @@ describe('catalogModuleGitlabDiscoveryEntityProvider', () => {
     await startTestBackend({
       extensionPoints: [[catalogProcessingExtensionPoint, extensionPoint]],
       services: [
-        [coreServices.config, config],
+        [coreServices.rootConfig, config],
         [coreServices.logger, getVoidLogger()],
         [coreServices.scheduler, scheduler],
       ],