+++ plugins/catalog-backend-module-gitlab/src/providers/GitlabDiscoveryEntityProvider.ts
@@ -155,6 +155,7 @@ export class GitlabDiscoveryEntityProvider implements EntityProvider {
     const projects = paginated<GitLabProject>(
       options => client.listProjects(options),
       {
+        archived: false,
         group: this.config.group,
         page: 1,
         per_page: 50,
@@ -173,7 +174,10 @@ export class GitlabDiscoveryEntityProvider implements EntityProvider {
 
       res.scanned++;
 
-      if (project.archived) {
+      if (
+        this.config.skipForkedRepos &&
+        project.hasOwnProperty('forked_from_project')
+      ) {
         continue;
       }