+++ plugins/linguist-backend/src/index.ts
@@ -20,6 +20,7 @@
  * @packageDocumentation
  */
 
+export * from './processor';
 export * from './service/router';
 export type { LinguistBackendApi } from './api';
 export { linguistPlugin } from './plugin';