+++ plugins/playlist/src/components/CreatePlaylistButton/CreatePlaylistButton.tsx
@@ -14,7 +14,12 @@
  * limitations under the License.
  */
 
-import { errorApiRef, useApi, useRouteRef } from '@backstage/core-plugin-api';
+import {
+  errorApiRef,
+  useApi,
+  useRouteRef,
+  alertApiRef,
+} from '@backstage/core-plugin-api';
 import { BackstageTheme } from '@backstage/theme';
 import { usePermission } from '@backstage/plugin-permission-react';
 import {
@@ -34,6 +39,7 @@ import { useTitle } from '../../hooks';
 export const CreatePlaylistButton = () => {
   const navigate = useNavigate();
   const errorApi = useApi(errorApiRef);
+  const alertApi = useApi(alertApiRef);
   const playlistApi = useApi(playlistApiRef);
   const playlistRoute = useRouteRef(playlistRouteRef);
   const [openDialog, setOpenDialog] = useState(false);
@@ -49,11 +55,16 @@ export const CreatePlaylistButton = () => {
       try {
         const playlistId = await playlistApi.createPlaylist(playlist);
         navigate(playlistRoute({ playlistId }));
+        alertApi.post({
+          message: `Added playlist '${playlist.name}'`,
+          severity: 'success',
+          display: 'transient',
+        });
       } catch (e) {
         errorApi.post(e);
       }
     },
-    [errorApi, navigate, playlistApi, playlistRoute],
+    [errorApi, navigate, playlistApi, playlistRoute, alertApi],
   );
 
   const singularTitle = useTitle({