+++ plugins/apache-airflow/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @backstage/plugin-apache-airflow
 
+## 0.2.14-next.1
+
+### Patch Changes
+
+- 12a8c94eda8d: Add package repository and homepage metadata
+- Updated dependencies
+  - @backstage/core-components@0.13.4-next.0
+  - @backstage/core-plugin-api@1.5.3
+
+## 0.2.14-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.13.4-next.0
+  - @backstage/core-plugin-api@1.5.3
+
 ## 0.2.13
 
 ### Patch Changes