+++ plugins/cicd-statistics/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @backstage/plugin-cicd-statistics
 
+## 0.1.23-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-catalog-react@1.8.1-next.1
+
+## 0.1.23-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-plugin-api@1.5.3
+  - @backstage/plugin-catalog-react@1.8.1-next.0
+  - @backstage/catalog-model@1.4.1
+
 ## 0.1.22
 
 ### Patch Changes