+++ plugins/git-release-manager/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @backstage/plugin-git-release-manager
 
+## 0.3.34-next.1
+
+### Patch Changes
+
+- 12a8c94eda8d: Add package repository and homepage metadata
+- Updated dependencies
+  - @backstage/integration@1.5.1
+  - @backstage/core-components@0.13.4-next.0
+  - @backstage/core-plugin-api@1.5.3
+  - @backstage/theme@0.4.1
+
+## 0.3.34-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.13.4-next.0
+  - @backstage/core-plugin-api@1.5.3
+  - @backstage/integration@1.5.1
+  - @backstage/theme@0.4.1
+
 ## 0.3.33
 
 ### Patch Changes