+++ plugins/graphql-backend/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @backstage/plugin-graphql-backend
 
+## 0.1.38-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.19.2-next.2
+
+## 0.1.38-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.19.2-next.1
+  - @backstage/config@1.0.8
+  - @backstage/plugin-catalog-graphql@0.3.22
+
+## 0.1.38-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.19.2-next.0
+  - @backstage/config@1.0.8
+  - @backstage/plugin-catalog-graphql@0.3.22
+
 ## 0.1.37
 
 ### Patch Changes