+++ plugins/newrelic/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-newrelic
 
+## 0.3.39-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.13.4-next.0
+  - @backstage/core-plugin-api@1.5.3
+  - @backstage/theme@0.4.1
+
 ## 0.3.38
 
 ### Patch Changes