+++ plugins/stack-overflow-backend/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @backstage/plugin-stack-overflow-backend
 
+## 0.2.4-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.19.2-next.2
+
+## 0.2.4-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.19.2-next.1
+  - @backstage/config@1.0.8
+  - @backstage/plugin-search-common@1.2.5
+
+## 0.2.4-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.19.2-next.0
+  - @backstage/config@1.0.8
+  - @backstage/plugin-search-common@1.2.5
+
 ## 0.2.3
 
 ### Patch Changes